Fixes #1450
